factor out now-duplicated generate_union_fix
These were only very nearly duplicates before, but after passing in the
`importer` they would have become exact duplicates.
