+++ crates/ruff_linter/src/rules/flake8_pyi/rules/redundant_numeric_union.rs
@@ -1,17 +1,14 @@
 use bitflags::bitflags;
 
-use anyhow::Result;
-
 use ruff_diagnostics::{Applicability, Diagnostic, Edit, Fix, FixAvailability, Violation};
 use ruff_macros::{derive_message_formats, ViolationMetadata};
-use ruff_python_ast::{
-    name::Name, AnyParameterRef, Expr, ExprBinOp, ExprContext, ExprName, ExprSubscript, ExprTuple,
-    Operator, Parameters, PythonVersion,
-};
+use ruff_python_ast::{AnyParameterRef, Expr, ExprBinOp, Operator, Parameters, PythonVersion};
 use ruff_python_semantic::analyze::typing::traverse_union;
 use ruff_text_size::{Ranged, TextRange};
 
 use crate::checkers::ast::Checker;
+
+use super::generate_union_fix;
 
 /// ## What it does
 /// Checks for parameter annotations that contain redundant unions between
@@ -157,12 +154,18 @@ fn check_annotation<'a>(checker: &Checker, annotation: &'a Expr) {
                 applicability,
             )),
             UnionKind::TypingUnion => {
-                let Ok(fix) =
-                    generate_union_fix(checker, necessary_nodes, annotation, applicability)
+                let Some(importer) = checker.typing_importer("Union", PythonVersion::lowest())
                 else {
                     return;
                 };
-                fix
+                generate_union_fix(
+                    checker.generator(),
+                    &importer,
+                    necessary_nodes,
+                    annotation,
+                    applicability,
+                )
+                .ok()
             }
         }
     };
@@ -261,43 +264,4 @@ fn generate_pep604_fix(
         Edit::range_replacement(checker.generator().expr(&new_expr), annotation.range()),
         applicability,
     )
-}
-
-/// Generate a [`Fix`] for two or more type expressions, e.g. `typing.Union[int, float, complex]`.
-fn generate_union_fix(
-    checker: &Checker,
-    nodes: Vec<&Expr>,
-    annotation: &Expr,
-    applicability: Applicability,
-) -> Result<Option<Fix>> {
-    debug_assert!(nodes.len() >= 2, "At least two nodes required");
-
-    // Request `typing.Union`
-    let Some(importer) = checker.typing_importer("Optional", PythonVersion::lowest()) else {
-        return Ok(None);
-    };
-    let (import_edit, binding) = importer.import(annotation.start())?;
-
-    // Construct the expression as `Subscript[typing.Union, Tuple[expr, [expr, ...]]]`
-    let new_expr = Expr::Subscript(ExprSubscript {
-        range: TextRange::default(),
-        value: Box::new(Expr::Name(ExprName {
-            id: Name::new(binding),
-            ctx: ExprContext::Store,
-            range: TextRange::default(),
-        })),
-        slice: Box::new(Expr::Tuple(ExprTuple {
-            elts: nodes.into_iter().cloned().collect(),
-            range: TextRange::default(),
-            ctx: ExprContext::Load,
-            parenthesized: false,
-        })),
-        ctx: ExprContext::Load,
-    });
-
-    Ok(Some(Fix::applicable_edits(
-        Edit::range_replacement(checker.generator().expr(&new_expr), annotation.range()),
-        [import_edit],
-        applicability,
-    )))
 }
